fix: split out validation into its own class
